embeddings/requirements.txt
torch
sentence-transformers
pandas
pyarrow
deltalake
s3fs
faiss-cpu
python-dotenv